refactor: handleRefInputMaxLength
